Move tests feature tests to support/util
