+++ tests/unit/compose.ts
@@ -1,37 +1,7 @@
 import * as registerSuite from 'intern!object';
 import * as assert from 'intern/chai!assert';
+import { hasToStringTag, hasConfigurableName } from '../support/util';
 import compose, { GenericClass, ComposeMixinDescriptor, getInitFunctionNames } from '../../src/compose';
-
-let _hasConfigurableName: boolean;
-
-/**
- * Detects if functions have configurable names, some browsers that are not 100% ES2015
- * compliant do not.
- */
-function hasConfigurableName(): boolean {
-	if (_hasConfigurableName !== undefined) {
-		return _hasConfigurableName;
-	}
-	const nameDescriptor = Object.getOwnPropertyDescriptor(function foo() {}, 'name');
-	if (nameDescriptor && !nameDescriptor.configurable) {
-		return _hasConfigurableName = false;
-	}
-	return _hasConfigurableName = true;
-}
-
-let _hasToStringTag: boolean;
-
-/**
- * Detects if the runtime environment supports specifying a Symbol.toStringTag
- */
-function hasToStringTag(): boolean {
-	if (_hasToStringTag !== undefined) {
-		return _hasToStringTag;
-	}
-	const a: any = {};
-	a[Symbol.toStringTag] = 'foo';
-	return _hasToStringTag = (a + '') === '[object foo]';
-}
 
 registerSuite({
 	name: 'lib/compose',
